PokeAPI - Listing Pokemon
In this build, listing all of the pokemon is now working. However there is a commented out IF statement that tells it to ignore ID 71. This was due to an error, where the data for ID 71 wasnt availaable and was causing errors. This has been resolved since, however i am keeping the IF statemet inside the code, commented out, just in case something similar to this occurs again.

The next goal is to filter by name, then to filter by stats. So filter a pokemon by the highest HP, Attack, Defence, Special Attack, Special Defence, Type, etc.