PokeAPI - New showcase and search fix
Showcase - Instead of being displayed in a table, the pokemon are now displayed in a grid. This is more visually appealing and looks a lot more clean.

Search Fix - I was formatting the Pokemons names into title case to make them more visually appealing and this was causing problems in the search when trying to search for a pokemon with all lowercase or all uppercase. I fixed it by having any string typed into the input box, converted to a lowercase string.