PokeAPI - New Layout
Now i have the types working properly, i have made the layout correct. A grid showing the pokemon, and at the side will be the place for filtering. So far search works and i will be adding filtering by type in my next commit. Clicking on a Pokemon in the grid routes to the pokemon details page, however i have not decided on how i want this to look.